Diciembre 15 de 2016
Subido por Miguel Angel Manrique Latorre
Se Adiciona la columna Clase Riesgo a la pantalla vigencias cargos.
Se crea la función para generar el reporte funciones_cargo.
Se crea el dialogo para adicionar una vigencia ARP en la página vigencias cargos

+++ designer/designer-ejb/src/java/Entidades/ClasesRiesgos.java
@@ -28,7 +28,7 @@ public class ClasesRiesgos implements Serializable {
     @Column(name = "DESCRIPCION")
     private String descripcion;
     @Column(name = "PORCENTAJE")
-    private BigInteger porcentaje;
+    private BigDecimal porcentaje;
 
     public ClasesRiesgos() {
     }
@@ -61,11 +61,11 @@ public class ClasesRiesgos implements Serializable {
         this.descripcion = descripcion;
     }
 
-    public BigInteger getPorcentaje() {
+    public BigDecimal getPorcentaje() {
         return porcentaje;
     }
 
-    public void setPorcentaje(BigInteger porcentaje) {
+    public void setPorcentaje(BigDecimal porcentaje) {
         this.porcentaje = porcentaje;
     }
